% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Algae dataset}
\usage{
clean(df)
}
\arguments{
\item{df}{The data frame to be cleaned}
}
\value{
The cleaned data frame
}
\description{
Takes in a data frame and cleans it properly for use in the algaeplot
}

