% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumNA.R
\name{sumNA}
\alias{sumNA}
\title{Sum of Vector Elements}
\usage{
sumNA(x)
}
\arguments{
\item{x}{A numeric vector to find the mean of}
}
\value{
The sum of the vector x
}
\description{
Takes in a numeric vector and returns the sum of it. If at least one entry is numeric, then the NA entries are treated as zeros. If no entry is a value, then it returns NA.
}
\examples{
x <- c(5,3,7,2,8,5,6,4,NA,NA,3,2,NA,4)
sumNA(x)
returns: 49
}

