% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sumZero.R
\name{sumZero}
\alias{sumZero}
\title{Sum of Vector Elements}
\usage{
sumZero(x)
}
\arguments{
\item{x}{A numeric vector to find the sum of}
}
\value{
The sum of the vector x
}
\description{
Takes in a numeric vector and returns the sum of it. If at least one entry is numeric, then the function returns 0. If no entry is a value, then it returns NA.
}
\examples{
x <- c(NA,2)
sumZero(x)
returns: 0
}

